+++ src/hvl/projectparmorel/modelrepair/ActionExtractor.java
@@ -1,4 +1,4 @@
-package hvl.projectparmorel.moderrepair;
+package hvl.projectparmorel.modelrepair;
 
 import java.lang.reflect.Method;
 import java.util.ArrayList;
